# BookBackend/api/openAPI.yaml
openapi: "3.0.3"
info:
  title: "BookBackend API"
  description: "BookBackend API"
  version: "1.0.0"
servers:
  - url: "https://BookBackend"
paths:
...
  /authors:
      get:
        tags:
            - "Authors"
        summary: "GET authors"
        operationId: "getAllAuthors"
        responses:
          "200":
            $ref: "#/components/responses/AuthorList"
          "400":
            $ref: "#/components/responses/BadRequest"
          "401":
            $ref: "#/components/responses/Unauthorized"
      post:
          tags:
              - "Authors"
          summary: "POST authors"
          operationId: "createAuthor"
          requestBody:
              $ref: "#/components/requestBodies/CreateAuthor"
          responses:
            "201":
              $ref: "#/components/responses/CreateAuthor"
            "400":
              $ref: "#/components/responses/BadRequest"
            "401":
              $ref: "#/components/responses/Unauthorized"
          security:
            - prod: [ ]
  /authors/{authorId}:
        get:
            tags:
                - "Authors"
            summary: "GET authors/{id}"
            operationId: "getAuthorById"
            parameters:
              - $ref: "#/components/parameters/authorId"
            responses:
              "200":
                $ref: "#/components/responses/Author"
              "400":
                $ref: "#/components/responses/BadRequest"
              "401":
                $ref: "#/components/responses/Unauthorized"
        put:
            tags:
                - "Authors"
            summary: "PUT authors/{id}"
            operationId: "updateAuthor"
            parameters:
              - $ref: "#/components/parameters/authorId"
            requestBody:
                  $ref: "#/components/requestBodies/UpdateAuthor"
            responses:
                "200":
                  $ref: "#/components/responses/Author"
                "400":
                  $ref: "#/components/responses/BadRequest"
                "401":
                  $ref: "#/components/responses/Unauthorized"
            security:
              - prod: [ ]
        delete:
            tags:
                - "Authors"
            summary: "DELETE authors/{id}"
            operationId: "deleteAuthor"
            parameters:
                - $ref: "#/components/parameters/authorId"
            responses:
                "200":
                    description: "OK"
                "400":
                  $ref: "#/components/responses/BadRequest"
                "401":
                  $ref: "#/components/responses/Unauthorized"
            security:
              - prod: [ ]
...
components:
  parameters:
...
    authorId:
        name: authorId
        in: path
        description: ID of author to return
        required: true
        schema:
            type: integer
            format: int64
...
  requestBodies:
...
    CreateAuthor:
      description: "Create author"
      required: true
      content:
          application/json:
            schema:
                $ref: "#/components/schemas/CreateAuthor"
    UpdateAuthor:
      description: "Update author"
      required: true
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/UpdateAuthor"
...
  responses:
...
    AuthorList:
        description: "OK"
        content:
            application/json:
              schema:
                  type: "array"
                  items:
                    $ref: "#/components/schemas/Author"
    Author:
      description: "OK"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Author"
    CreateAuthor:
      description: "Created"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Author"
...
    BadRequest:
        description: "Bad Request"
        content:
            application/json:
              schema:
                  $ref: "#/components/schemas/Error"
    Unauthorized:
        description: "Unauthorized"
        content:
            application/json:
              schema:
                  $ref: "#/components/schemas/Error"
  schemas:
...
    Author:
        type: "object"
        properties:
            id:
              type: "integer"
              format: "int64"
            name:
              type: "string"
            surname:
              type: "string"
    CreateAuthor:
        type: "object"
        properties:
            name:
              type: "string"
            surname:
              type: "string"
    UpdateAuthor:
        type: "object"
        properties:
            name:
              type: "string"
            surname:
              type: "string"
...
    Error:
        type: "object"
        properties:
            message:
              type: "string"
              example: "Bad Request"
  securitySchemes:
    prod:
      scheme: bearer
      type: http
      bearerFormat: JWT